requirements.txt
# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0

# Vector DB and embeddings - versions compatibles
chromadb==0.4.15
 huggingface_hub==0.10.1
sentence-transformers==2.2.2
transformers==4.30.0
torch==2.7.0
huggingface-hub==0.16.4

# Utilities
numpy==1.26.0
requests==2.31.0
aiofiles==23.2.1

# Optional - for production
openai==1.3.7
